flowgrind: some work on pcap support, still not working though
git-svn-id: svn+ssh://svn.umic-mesh.net/umic-mesh/projects/flowgrind/trunk@6548 3d4b872d-e00e-0410-a2ce-a5c2cda4747c

+++ src/common.h
@@ -160,8 +160,8 @@ struct _report
 	unsigned long long bytes_read;
 	unsigned long long bytes_written;
 #else
-	unsigned int bytes_read;
-	unsigned int bytes_written;
+	long bytes_read;
+	long bytes_written;
 #endif 
 	unsigned int request_blocks_read;
 	unsigned int request_blocks_written;
